convert tag-explorer to new API
